requirements.txt
amqp==5.2.0
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
blinker==1.5
Brlapi==0.8.4
Brotli==1.1.0
celery==5.2.7
certifi==2022.9.24
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==38.0.4
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.3.2
Deprecated==1.2.14
distlib==0.3.9
distro==1.8.0
distro-info==1.5+deb12u1
Django==4.2
django-allauth==65.0.2
django-ckeditor==6.7.1
django-cleanup==9.0.0
django-crispy-forms==1.9.2
django-js-asset==2.2.0
djangorestframework==3.14.0
filelock==3.16.1
fonttools==4.50.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.0
idna==3.3
kombu==5.3.5
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
louis==3.24.0
Mako==1.2.4.dev0
MarkupSafe==2.1.2
mysqlclient==1.4.6
oauthlib==3.2.2
package-name==0.1
packaging==23.2
Pillow==10.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.43
psycopg==3.1.17
psycopg2-binary==2.9.3
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydyf==0.9.0
PyGObject==3.42.2
PyJWT==2.6.0
pyparsing==3.0.9
pyphen==0.14.0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.6.0
python-barcode==0.15.1
python-bidi==0.6.3
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
python-telegram-bot==21.1.1
pytz==2023.3.post1
pyxdg==0.28
redis==4.3.4
reportbug==12.0.0
requests==2.28.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
telegram==0.0.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.12
vine==5.1.0
virtualenv==20.26.6
wadllib==1.3.6
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
whitenoise==4.1
wrapt==1.16.0
xdg==5
zopfli==0.2.3
